repair rev 10000.  This math seems to work a lot better
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@10006 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ mount_slash/mount_slash_int.c
@@ -1660,7 +1660,7 @@ msl_io(struct msl_fhent *mfh, char *buf, size_t size, off_t off, enum rw rw)
 		/* The client cache is operating on pages (ie 32k) so 
 		 *   any short read must be caught here.
 		 */
-		if (fsz < (size + off))
+		if (fsz < (ssize_t)(size + off))
 			rc = (size - (size + off - fsz));
 		else
 			rc = size;
